refactors mws monitoring models and business logic
